requirements.txt
streamlit
matplotlib
Pillow
PyMuPDF
